feat: rename relations v2
