+++ packages/substream/sink/write-edits/get-stale-entities-from-relations.ts
